// src/MapMesh.d.ts
export interface MapMeshOptions {
    /**
     * Texture containing the sub textures referenced in the terrain texture atlas.
     */
    terrainAtlasTexture: Texture;
    /**
     * Texture atlas describing the sub textures for each terrain type.
     */
    terrainAtlas: TextureAtlas;
    /**
     * Texture with blend masks for transitions between terrains
     */
    transitionTexture: Texture;
    /**
     * River tile atlas texture containing parts for each possible river variation.
     * Use /tools/river-atlas.py to generate.
     */
    riverAtlasTexture: Texture;
    /**
     * Coast tile atlas texture containg parts for each possible coast variation.
     * Use /tools/coast-atlas.py to generate.
     */
    coastAtlasTexture: Texture;
    /**
     * The texture used for undiscovered tiles
     */
    undiscoveredTexture: Texture;
    /**
     * Normal map for hills
     */
    hillsNormalTexture: Texture;
    /**
     * Diffuse map for tree sprites
     */
    treeSpritesheet: Texture;
    /**
     * Number of horizontal and vertical spritesheet subdivisions
     */
    treeSpritesheetSubdivisions: number;
    /**
     * Default 1.0
     */
    treeSize?: number;
    /**
     * Parts of the tree sprite whose opacity is lower than this value will not be rendered,
     * i.e. the transparent background. Valid values are between 0.0 and 1.0. Default is 0.2.
     */
    treeAlphaTest?: number;
    /**
     * Default 50
     */
    treesPerForest?: number;
    /**
     * Options per tree index to vary individual tree types.
     */
    treeOptions?: {
        /**
         * Tree size scale (1.0 by default)
         */
        scale?: number;
        treesPerForest: number;
    }[];
    /**
     * Overall scale of the geometry. Default 1.0
     */
    scale?: number;
    /**
     * GLSL code of the land fragment shader. For default see /src/shaders/land.fragment.ts.
     */
    landFragmentShader?: string;
    /**
     * GLSL code of the land vertex shader. For default see /src/shaders/land.vertex.ts.
     */
    landVertexShader?: string;
    /**
     * GLSL code of the mountain fragment shader. For default see /src/shaders/mountains.fragment.ts.
     */
    mountainsFragmentShader?: string;
    /**
     * GLSL code of the mountain vertex shader. For default see /src/shaders/mountains.vertex.ts.
     */
    mountainsVertexShader?: string;
    /**
     * Color of the hex grid
     * Default: 0xffffff (white)
     */
    gridColor?: Color;
    /**
     * Width of the grid lines as a normalized scaling factor relative to the area of the tile.
     * Default: 0.02 (2%)
     */
    gridWidth?: number;
    /**
     * Opacity between 0.0 (invisible) and 1.0 (opaque).
     * Default value: 0.33 (33%)
     */
    gridOpacity?: number;
}